requirements.txt
alembic==1.6.5
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
b4==0.8.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==3.3.0
blis==0.7.5
Brotli==1.0.9
bs4==0.0.1
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
chart-studio==1.1.0
click==8.0.3
cycler==0.11.0
cymem==2.0.6
dash==1.21.0
dash-bootstrap-components==0.13.0
dash-core-components==1.17.1
dash-html-components==1.1.4
dash-table==4.12.0
decorator==4.4.2
defusedxml==0.7.0
distlib==0.3.1
dkimpy==1.0.5
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.3
entrypoints==0.3
filelock==3.3.2
Flask==2.0.2
Flask-Bootstrap==3.3.7.1
Flask-Compress==1.10.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
future==0.18.2
greenlet==1.1.1
gunicorn==20.1.0
huggingface-hub==0.1.2
idna==3.3
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.3
joblib==1.1.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.2
lab==6.3
langcodes==3.2.1
llvmlite==0.37.0
lxml==4.6.4
Mako==1.1.4
MarkupSafe==2.0.1
matplotlib==3.4.3
mistune==0.8.4
murmurhash==1.0.6
mysql-connector==2.2.9
mysql-connector-python==8.0.23
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
nltk==3.6.5
notebook==6.2.0
numba==0.54.1
numpy==1.20.3
packaging==21.2
pandas==1.3.4
pandocfilters==1.4.3
parso==0.8.1
patatt==0.4.9
pathy==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.1.0
preshed==3.0.6
procfile==0.1.0
prometheus-client==0.9.0
prompt-toolkit==3.0.16
protobuf==3.15.5
ptyprocess==0.7.0
pycparser==2.21
pydantic==1.8.2
Pygments==2.8.0
PyNaCl==1.4.0
pynndescent==0.5.5
pyparsing==2.4.7
pyrsistent==0.17.3
pyspellchecker==0.6.2
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2021.3
PyYAML==6.0
pyzmq==22.0.3
qtconsole==5.0.2
QtPy==1.9.0
regex==2021.11.10
requests==2.26.0
retrying==1.3.3
sacremoses==0.0.46
scikit-learn==1.0.1
scipy==1.7.2
seaborn==0.11.1
Send2Trash==1.5.0
sentence-transformers==2.1.0
sentencepiece==0.1.96
simplejson==3.17.2
six==1.16.0
sklearn==0.0
smart-open==5.2.1
soupsieve==2.3.1
spacy==3.2.0
spacy-legacy==3.0.8
spacy-loggers==1.0.1
SQLAlchemy==1.4.22
srsly==2.4.2
tenacity==8.0.1
terminado==0.9.2
testpath==0.4.4
thinc==8.0.13
threadpoolctl==3.0.0
tokenizers==0.10.3
torch==1.10.0
torchvision==0.11.1
tornado==6.1
tqdm==4.62.3
traitlets==5.0.5
transformers==4.12.3
txt2tags==3.7
typer==0.4.0
typing-extensions==3.10.0.2
umap==0.1.1
umap-learn==0.5.2
urllib3==1.26.7
virtualenv==20.4.2
visitor==0.1.3
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.1
WTForms==2.3.3
